extract cgrecord and various helper methods
